enrollment-app/enrollment-function-app/requirements.txt
azure-functions
pytest>=7.4.0
pytest-cov>=4.1.0
